reuse collection utility functions
