+++ src/collect_into/mod.rs
@@ -1,6 +1,7 @@
 mod fixed_vec;
 mod par_collect_into;
 mod split_vec;
+pub(crate) mod utils;
 mod vec;
 
 pub use par_collect_into::ParCollectInto;
