+++ src/collect_into/split_vec.rs
@@ -1,9 +1,11 @@
 use super::par_collect_into::ParCollectIntoCore;
 use crate::{
+    collect_into::utils::split_vec_reserve,
     computations::{M, Values, X, Xfx},
     runner::ParallelRunner,
 };
 use orx_concurrent_iter::ConcurrentIter;
+#[cfg(test)]
 use orx_pinned_vec::PinnedVec;
 use orx_split_vec::{GrowthWithConstantTimeAccess, PseudoDefault, SplitVec};
 
@@ -17,7 +19,7 @@ where
 
     fn empty(iter_len: Option<usize>) -> Self {
         let mut vec = Self::pseudo_default();
-        reserve(&mut vec, iter_len);
+        split_vec_reserve(&mut vec, iter_len);
         vec
     }
 
@@ -27,7 +29,7 @@ where
         I: ConcurrentIter,
         M1: Fn(I::Item) -> O + Send + Sync,
     {
-        reserve(&mut self, m.par_len());
+        split_vec_reserve(&mut self, m.par_len());
         let (_num_spawned, pinned_vec) = m.collect_into::<R, _>(self);
         pinned_vec
     }
@@ -40,7 +42,7 @@ where
         Vo::Item: Send + Sync,
         M1: Fn(I::Item) -> Vo + Send + Sync,
     {
-        reserve(&mut self, x.par_len());
+        split_vec_reserve(&mut self, x.par_len());
         let (_num_spawned, pinned_vec) = x.collect_into::<R, _>(self);
         pinned_vec
     }
@@ -56,7 +58,7 @@ where
         F: Fn(&Vt::Item) -> bool + Send + Sync,
         M2: Fn(Vt::Item) -> Vo + Send + Sync,
     {
-        reserve(&mut self, xfx.par_len());
+        split_vec_reserve(&mut self, xfx.par_len());
         let (_num_spawned, pinned_vec) = xfx.collect_into::<R, _>(self);
         pinned_vec
     }
@@ -67,17 +69,4 @@ where
     fn length(&self) -> usize {
         self.len()
     }
-}
-
-fn reserve<T, G: GrowthWithConstantTimeAccess>(
-    split_vec: &mut SplitVec<T, G>,
-    len_to_extend: Option<usize>,
-) {
-    match len_to_extend {
-        None => {
-            let capacity_bound = split_vec.capacity_bound();
-            split_vec.reserve_maximum_concurrent_capacity(capacity_bound)
-        }
-        Some(len) => split_vec.reserve_maximum_concurrent_capacity(split_vec.len() + len),
-    };
 }
