+++ src/using/collect_into/split_vec.rs
@@ -1,9 +1,9 @@
+use crate::collect_into::utils::split_vec_reserve;
 use crate::using::Using;
 use crate::using::collect_into::u_par_collect_into::UParCollectIntoCore;
 use crate::using::computations::{UM, UX, UXfx};
 use crate::{computations::Values, runner::ParallelRunner};
 use orx_concurrent_iter::ConcurrentIter;
-use orx_pinned_vec::PinnedVec;
 use orx_split_vec::{GrowthWithConstantTimeAccess, PseudoDefault, SplitVec};
 
 impl<O, G> UParCollectIntoCore<O> for SplitVec<O, G>
@@ -19,7 +19,7 @@ where
         I: ConcurrentIter,
         M1: Fn(&mut U::Item, I::Item) -> O + Send + Sync,
     {
-        reserve(&mut self, m.par_len());
+        split_vec_reserve(&mut self, m.par_len());
         let (_num_spawned, pinned_vec) = m.collect_into::<R, _>(self);
         pinned_vec
     }
@@ -33,7 +33,7 @@ where
         Vo::Item: Send + Sync,
         M1: Fn(&mut U::Item, I::Item) -> Vo + Send + Sync,
     {
-        reserve(&mut self, x.par_len());
+        split_vec_reserve(&mut self, x.par_len());
         let (_num_spawned, pinned_vec) = x.collect_into::<R, _>(self);
         pinned_vec
     }
@@ -53,21 +53,8 @@ where
         F: Fn(&mut U::Item, &Vt::Item) -> bool + Send + Sync,
         M2: Fn(&mut U::Item, Vt::Item) -> Vo + Send + Sync,
     {
-        reserve(&mut self, xfx.par_len());
+        split_vec_reserve(&mut self, xfx.par_len());
         let (_num_spawned, pinned_vec) = xfx.collect_into::<R, _>(self);
         pinned_vec
     }
-}
-
-fn reserve<T, G: GrowthWithConstantTimeAccess>(
-    split_vec: &mut SplitVec<T, G>,
-    len_to_extend: Option<usize>,
-) {
-    match len_to_extend {
-        None => {
-            let capacity_bound = split_vec.capacity_bound();
-            split_vec.reserve_maximum_concurrent_capacity(capacity_bound)
-        }
-        Some(len) => split_vec.reserve_maximum_concurrent_capacity(split_vec.len() + len),
-    };
 }
