+++ src/using/collect_into/vec.rs
@@ -1,3 +1,4 @@
+use crate::collect_into::utils::extend_vec_from_split;
 use crate::computations::Values;
 use crate::runner::ParallelRunner;
 use crate::using::Using;
@@ -5,8 +6,7 @@ use crate::using::collect_into::u_par_collect_into::UParCollectIntoCore;
 use crate::using::computations::{UM, UX, UXfx};
 use orx_concurrent_iter::ConcurrentIter;
 use orx_fixed_vec::FixedVec;
-use orx_pinned_vec::PinnedVec;
-use orx_split_vec::{GrowthWithConstantTimeAccess, SplitVec};
+use orx_split_vec::SplitVec;
 
 impl<O> UParCollectIntoCore<O> for Vec<O>
 where
@@ -23,7 +23,7 @@ where
             None => {
                 let split_vec = SplitVec::with_doubling_growth_and_max_concurrent_capacity();
                 let split_vec = split_vec.u_m_collect_into::<R, _, _, _>(m);
-                extend_from_split(self, split_vec)
+                extend_vec_from_split(self, split_vec)
             }
             Some(len) => {
                 self.reserve(len);
@@ -45,7 +45,7 @@ where
     {
         let split_vec = SplitVec::with_doubling_growth_and_max_concurrent_capacity();
         let split_vec = split_vec.u_x_collect_into::<R, _, _, _, _>(x);
-        extend_from_split(self, split_vec)
+        extend_vec_from_split(self, split_vec)
     }
 
     fn u_xfx_collect_into<R, U, I, Vt, Vo, M1, F, M2>(
@@ -65,20 +65,6 @@ where
     {
         let split_vec = SplitVec::with_doubling_growth_and_max_concurrent_capacity();
         let split_vec = split_vec.u_xfx_collect_into::<R, _, _, _, _, _, _, _>(xfx);
-        extend_from_split(self, split_vec)
-    }
-}
-
-fn extend_from_split<T, G>(mut initial_vec: Vec<T>, collected_split_vec: SplitVec<T, G>) -> Vec<T>
-where
-    G: GrowthWithConstantTimeAccess,
-{
-    match initial_vec.len() {
-        0 => collected_split_vec.to_vec(),
-        _ => {
-            initial_vec.reserve(collected_split_vec.len());
-            initial_vec.extend(collected_split_vec);
-            initial_vec
-        }
+        extend_vec_from_split(self, split_vec)
     }
 }
